Improve the marking of sent events
In generate_output(), we were marking events as sent by changing the XML
tree and saving it for each form event sent. This change replaces it
with a "sent event index" which simply writes out the form event to a
file and loads it back in if we use the --resume switch.

I've seen a 12% speed improvement during testing.