Update warning about unsent events
Updated the warning message to check the log file first. Also, after
discussing it with some of the other developers, I decided to make the
sent_events file more human-readable. In doing so, I implemented a
faster approach to persisting the list of sent events.  Instead of
pickling the list each time, the latest event simply has to be appended
to the file.